requirements.txt
google-genai
python-dotenv
matplotlib
streamlit
altair
pandas
supabase